Move loading graph logic into utils
